# perf(polys): make ZZ_I and QQ_I faster